+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Automatically set the correct class when generating known components
 - Add generator `components` that is able to mass-produce the most used components
 - Make fields point to the correct `model_class` in case of STI
+- Support hidden Anchormodel fields
 
 # 0.0.4